+++ src/compiler/expressionConfigTypes.ts
@@ -1,0 +1,16 @@
+import { BigNumberish, BytesLike } from 'ethers';
+
+/**
+ * @public
+ * Type of valid parsed expression, i.e. compiled bytes 
+ */
+export type ExpressionConfig = {
+    /**
+     * Sources verbatim.
+     */
+    sources: BytesLike[];
+    /**
+     * Constants verbatim.
+     */
+    constants: BigNumberish[];
+}
