+++ src/index.ts
@@ -1,8 +1,9 @@
-export * from './compiler/rainDecompiler';
 export * from './compiler/rainCompiler';
+export * from './compiler/rainDecompiler';
+export * from './compiler/expressionConfigTypes';
 
 export * from './parser/rainParser';
+export * from './parser/opMetaTypes';
 export * from './parser/rainParserTypes';
 
-export * from './types';
 export * from './utils';
